move splitting multidoc to util
